Originally, statsdb compared the mtime of a file to the mtime it had stored in the DB to see whether the file changed.
On some file systems, mtime only has one second resolution, which means
that if a file changes quickly enough after the statsdb has stored its
mtime, the statsdb won't realize that the file has changed, and it will
therefore supply incorrect stats about the file.

In order for this to be ameliorated, we now also take the file size into
account and we store a value which contains mtime bits as well as file
size bits into the database. The only files which can now slip past the
system are files which are modified soon enough and which retain their
sizes. And honestly, such files, if they are valid translation files,
are so unlikely to contain significant changes which maintain the file
size that they can be ignored.

svn path=/src/trunk/; revision=7507