A number of methods in UnitChecker were called very, very frequently and this contributed a lot to the running time of checks.
Now, the results of a number of these methods are memoized (by the
decorator cache_results). The cached results are cleared every time
that TranslationChecker.run_test is called.

On a small OpenOffice project, this has sliced the computation time
of stats quite a lot and it should never result in a slow-down.

Note that the decorator is deliberately specialized to accept only
single argument methods. This is to avoid the runtime cost of unpacking
arbitrary numbers of parameters.

svn path=/src/trunk/; revision=7521