Removed the HORRIBLE hack whereby one could pass parameters to the stats code in order to get it not to throw an exception.
I ought to have my coding licence revoked. But at least until then, I
can try to restore my karma by restoring the code to a better state.

So anyway, if you call one of the stats functions on a non-existant or
illegal file, it's your responsibility to deal with the exceptions
that might pop out.

svn path=/src/trunk/; revision=7796