Preempt missing source and existing target errors
The .ini files Mozilla uses sometimes points to non-existant
directories. We preempt that and ignore them.
As we run the command twice we also might be writing onto an
existing target.  We ignore those also.
The verbosity is handles for each of those.
For real errors we output the error messages as a last resort.