Optimise unquoting and unescaping.
This is a new PO specific implementation of the over
general quote.extractwithoutquotes(). This results in
a ~60% reduction in time for unquoting and unescaping.
This should improve the speed of anything that calls
(pypo) pounit.source and .target.

extractpoline() is kept to allow further tests of
correctness, and in case of external users.