Be explicit about the merging criterion to use.
This adds support for a new custom header, 'X-Merge-On', that specifies
the type of matching that will be used when merging units. For now it
accepts two values ('location', 'id'), being 'id' the default when the
custom header is absent.

Prior to this, the 'X-Accelerator-Marker' was used doing wild guesses
about the matching criterion to use when merging. This has lead to
confusion and issues with PO files generated at LibreOffice for instance
(bug 2587), where the desired behavior was merging based on IDs
(combination of msgid and msgctxt) rather than on location, and this
resulted in lots of units being marked as fuzzy.