Handling Mozilla plural forms in properties files
Adds support for converting between Mozilla style and Gettext plural
forms.

Mozilla plurals is just a hack slapped on top of existing properties
files with no robust way to detect them, so we use a heuristic that
seems to catch the largest number of strings with no false positives,
but there are some false negatives.

No attempt is made to validate the forms.

Part of bug 329