Cleanup: remove typecheck library
Typechecking was a nice idea and found some bugs in our code.

The library is not maintained; https://pypi.python.org/pypi/typecheck
with last release for Python 2.4 and 2.5. It works only because we drag
it along with us and it probably won't work in Python 3 and we certainly
don't want to be responsible for fixing it.

I believe Python 3 could potentially have a proper implementation, see
http://www.python.org/dev/peps/pep-3107/.

This removes the parts from misc/, all imports of misc.typecheck, all
@accepts and @returns decorators as well as special uses of IsOneOf and
Class.