docs: cPO merging is mostly done
While there are still bugs in cPO there are better ways to track this.
It is too expensive in terms of time to review and port any issues to
bugzilla so not bothering to port any of these. We'll do them when we
find them and find the time to fix them.