pypo: correctly mark fuzzy units
I'm not sure why this is needed, I would have thought setting self.STATE
would be enough so that when we serialise the PO we get the correct
fuzzy markers.

This fixes a poswap bug were a target language having a fuzzy unit did
not have the fuzzy marker carried over to the combined file.  Thus the
combined file would contain a translated unit that was fuzzy in the
original language.