idml2po: Record in unit location the file within the IDML it was extracted from
This records a second location on each unit, besides the other location
used to keep the xpath.

This is necessary because the files we are extracting from the IDML packages
have the same root tag, where the ODF packages have a different root tag
for each of the translatable files we extract.

This change doesn't require any alterations in the ODF converters.