multistring: fix list coercion to text
At the moment code such as `'%s' % [multistring(u'føø')]` fails: the
representation of `multistring` is used when coercing the list to text, however
there are decoding errors.

This commit adjusts the representation of `multistring`s so that a) coercing
lists containing `multistring`s is fixed, and b) the representation allows to
re-create the original object.

On b), this is what the official docs on `repr()` suggest at
https://docs.python.org/2/library/functions.html#repr:
>  For many types, this function makes an attempt to return a string that would
>  yield an object with the same value when passed to eval(), otherwise the
>  representation is a string enclosed in angle brackets that contains the name
>  of the type of the object together with additional information often
>  including the name and address of the object.