MozillaChecker's accelerators checks for accelerator absence
Some languages, mainly Indic, do not use accelerators, so instead of
checking that the accelerator is present it checks for its absence.

By altering the accelerators check behavior there is no need to
disable it anymore for some languages. Anyway the changes introduced
to be able to disable some check for some language in a specific
checker are kept since it is possible it is needed in the future.

Note behavior is altered based on the script used for the language.

Fixes #3479.