Add option to podebug to avoid rewriting placeholders
This change adds the optional --preserveplaceholders flag to podebug, which when enabled ensures that "unicode" and "flipped" rewrites are not applied to sub-strings identified as placeholders. In this way, placeholders remain unchanged in the output so that systems working with the rewritten strings can still inject data at run-time as normal.