Docs: use Sphinx internal mocking approach
Used for everything except two dependencies where we need to perform
version checking, for those use the old approach.

This is needed as things like xml are used by Sphinx itself and if a
builder requires it then it can't as we've mocked it.  The internal
Sphinx approach mocks it for the docs, but stll has the real lib
available to Sphinx itself.