Avoid potential Xapian database corruption
Xapian's flintlock is not a lock file whose presence indicates that the
lock is held, but rather a file which is locked using fcntl() locking.
Such locks are automatically released by the OS when the process holding
them terminates, so there's no such thing as a "stale lock", and the
flintlock file should always exist.

If you delete the lockfile and there's an existing writer holding the
lock, it'll still believe it holds the lock but now a second writer can
now also get the lock - if both writers try to update the database, the
result will be database corruption.

Fixes #3581.