Base store: Use first unit in locationindex if several units have same location
Note this impacts all the formats. But is needed to be able to output duplicate
units in properties when syncing back with po2moz or po2prop.

The reasons to not implement using lists to track all the units with the same
location are:

- Since locationindex is mostly used to perform operations on two stores, for
  example in converters, we would have to rely on both stores having the units
  with the same location to be in the same order in order to reliably match the
  right units. This cannot be ensured with current implementation.
- It involves a non-trivial refactoring of several converters and tools in order
  to handle retrieval of the right units from the location's unit list in
  locationindex index. This is mostly not a good idea since part of the code
  needing refactoring is not covered by any testing at all.

In short, current approach of keeping a reference to the first matching unit just
allows to output duplicate units when syncing back with po2prop, even if the unit
is repeated as much times as the location is duplicated.