ResX: fix indent of the </data> element
roundtripping a .NET Resource file would cause the closing </data> tag
to be indented by four spaces when there is no <comment>, while the
original file only has two.

however, when there is a <comment> tag, it still needs to indent by
four spaces to make its alignment correct.

this causes a lot of noise in a VCS that could be avoided