Add Kabyle language
Adding kabyle entry for the languages dictionary.
Please note that we two code ara affected on CLDR: kab and kab_DZ
We are working on kab code in the CLDR project version 32 (now on submission).
The two codes are also considered on Mozilla Firefox I localized. The priority is for kab.
We used kab as is suggested on iso 639-2 for language codes.
______
This update is for wablate. We are going to localize OpenSuse, so we need this change.