Use splitlines instead of implicit iterating on file
When the file is open in binary mode (what usually is), the implicit
iterator works only on platform native newlines. For example that means
that parser would fail to parse po files with unix newlines on Windows.

On the other side the splitlines uses universal newlines handling
correctly any newlines on any platform.

Fixes #3751