Android: Avoid using Babel for plural tags
These seem to be broken as they are missing the 'other' tag. The problem
is that not all languages use 'other' tag, so adding it back requires
CLDR knowledge. In this case it's better to have full tags defined
inside translate-toolkit than relying on broken library and fixing the
data.

See following issue trackers for more details:

* https://github.com/WeblateOrg/weblate/issues/1550
* https://github.com/python-babel/babel/issues/495