Catkeys: merge header and fingerprint
The catkeys format now has support for fingerprint calculation.
This means that if the tools change the contents of the file (like
remove or change strings), the header will now be correct and the file
will be parseable on Haiku.

When merging with pot2po, it will use the template header to determine
the language of the output format.