Use backports.csv module on Python 2
This allows to drop many hacks on csv handling as Python 2 and 3
versions now behave same. Also this fixes several corner cases in
handling CSV files (eg. BOM charsets were wrongly saved on Python 2 as
that did lead to having BOM in every field instead of just at the
beginning of the file).