Android: Improve XML handling
The root element can have a namespace and that broke the string
replacements. Instead of detecting the content, force the XML to include
proper closing tag (instead of self-closing one). This way the text
replacement is safe and does not break on additional attributes defined
on the root element.

Fixes https://github.com/WeblateOrg/weblate/discussions/5657