fluent: Refactor tests and add more coverage.
We refactor the fluent tests by reproducing the common test pattern in
`basic_test`. We also automatically test for any unspecified FluentUnit
properties in `assert_units` to make sure a unit does not have a value
when none is expected.

We also add some coverage for previously untested features:

+ Term Attributes.
+ FunctionReferences.
+ Literals.
+ Multi-line values for Terms, Messages and Attributes.
+ Whitespace.
+ Handling of special characters.
+ FluentUnits with an empty source.
+ Expected serialize failures.

And we add more test coverage for other features.