po2rc: Correctly escape double-quotes embedded in a string in a RC file
See https://learn.microsoft.com/en-us/windows/win32/menurc/stringtable-resource

> To embed quotes in the string, use the following sequence: "".

Because we may have some already translated strings which contain a
literal "" (i.e. \"\" in a .po file) to work-around po2rc not doing this
correctly so far, for backwards compatibility look around the " and only
escape it if it isn't already part of a "" escape.

Future work: This doesn't do anything for untranslated strings. That's
hard because those are passed through verbatim, including their existing
formatting and linebreaks.