feat(html): add ID as fallback for custom msgctxt
Follow-up to #5867.

Suggested in #4369. This automatically adds a fair bit of stability without needing to add a custom data-translate-context attribute.

It does have the unfortunate side-effect of not being backwards compatible with the existing msgctxt, but that's about the same as the current situation when changing the source file with a more positive end result (i.e., much less chance of such a mismatch happening in the future).